Added delete/edit lock for expenses. Added api key per user. Added manual...

See merge request expense_budget_tracker/expense_tracker!12

+++ backend/expense_tracker_backend/src/main/kotlin/mk/sorsix/com/expense_tracker_backend/service/ExpenseService.kt
@@ -64,6 +64,10 @@ class ExpenseService(
             return UpdateExpenseResult.NotOwner
         }
 
+        if (!existing.active) {
+            return UpdateExpenseResult.ExpenseLocked
+        }
+
         val foundCategory = categoryService.findCategoryById(request.categoryId)
             ?: return UpdateExpenseResult.CategoryNotFound
 
@@ -90,6 +94,10 @@ class ExpenseService(
 
         if (existing.user.id != user.id) {
             return DeleteExpenseResult.NotOwner
+        }
+
+        if (!existing.active) {
+            return DeleteExpenseResult.ExpenseLocked
         }
 
         expenseRepository.delete(existing)
